+++ src/dispatching/dialogue/transition.rs
@@ -1,20 +1,19 @@
-use crate::{
-    dispatching::{dialogue::DialogueStage, UpdateWithCx},
-    types::Message,
-};
+use crate::dispatching::{dialogue::DialogueStage, UpdateWithCx};
 use futures::future::BoxFuture;
+use teloxide_core::types::Message;
 
 /// Represents a transition function of a dialogue FSM.
 pub trait Transition: Sized {
     type Aux;
     type Error;
+    type Requester;
 
     /// Turns itself into another state, depending on the input message.
     ///
     /// `aux` will be passed to each subtransition function.
     fn react(
         self,
-        cx: TransitionIn,
+        cx: TransitionIn<Self::Requester>,
         aux: Self::Aux,
     ) -> BoxFuture<'static, TransitionOut<Self, Self::Error>>;
 }
@@ -29,6 +28,7 @@ where
     type Aux;
     type Dialogue;
     type Error;
+    type Requester;
 
     /// Turns itself into another state, depending on the input message.
     ///
@@ -36,7 +36,7 @@ where
     /// message's text.
     fn react(
         self,
-        cx: TransitionIn,
+        cx: TransitionIn<Self::Requester>,
         aux: Self::Aux,
     ) -> BoxFuture<'static, TransitionOut<Self::Dialogue, Self::Error>>;
 }
@@ -44,6 +44,7 @@ where
 /// A type returned from a FSM subtransition function.
 ///
 /// Now it is used only inside `#[teloxide(subtransition)]` for type inference.
+#[doc(hidden)]
 pub trait SubtransitionOutputType {
     type Output;
     type Error;
@@ -55,7 +56,7 @@ impl<D, E> SubtransitionOutputType for TransitionOut<D, E> {
 }
 
 /// An input passed into a FSM (sub)transition function.
-pub type TransitionIn = UpdateWithCx<Message>;
+pub type TransitionIn<R> = UpdateWithCx<R, Message>;
 
 /// A type returned from a FSM (sub)transition function.
 pub type TransitionOut<D, E = crate::RequestError> = Result<DialogueStage<D>, E>;