+++ src/dispatching/dispatcher.rs
@@ -3,34 +3,37 @@ use crate::{
         update_listeners, update_listeners::UpdateListener, DispatcherHandler, UpdateWithCx,
     },
     error_handlers::{ErrorHandler, LoggingErrorHandler},
+};
+use futures::StreamExt;
+use std::{fmt::Debug, sync::Arc};
+use teloxide_core::{
+    requests::Requester,
     types::{
         CallbackQuery, ChosenInlineResult, InlineQuery, Message, Poll, PollAnswer,
         PreCheckoutQuery, ShippingQuery, UpdateKind,
     },
-    Bot,
 };
-use futures::StreamExt;
-use std::{fmt::Debug, sync::Arc};
 use tokio::sync::mpsc;
 
-type Tx<Upd> = Option<mpsc::UnboundedSender<UpdateWithCx<Upd>>>;
+type Tx<Upd, R> = Option<mpsc::UnboundedSender<UpdateWithCx<Upd, R>>>;
 
 #[macro_use]
 mod macros {
     /// Pushes an update to a queue.
     macro_rules! send {
-        ($bot:expr, $tx:expr, $update:expr, $variant:expr) => {
-            send($bot, $tx, $update, stringify!($variant));
+        ($requester:expr, $tx:expr, $update:expr, $variant:expr) => {
+            send($requester, $tx, $update, stringify!($variant));
         };
     }
 }
 
-fn send<'a, Upd>(bot: &'a Bot, tx: &'a Tx<Upd>, update: Upd, variant: &'static str)
+fn send<'a, R, Upd>(requester: &'a R, tx: &'a Tx<R, Upd>, update: Upd, variant: &'static str)
 where
     Upd: Debug,
+    R: Requester + Clone,
 {
     if let Some(tx) = tx {
-        if let Err(error) = tx.send(UpdateWithCx { bot: bot.clone(), update }) {
+        if let Err(error) = tx.send(UpdateWithCx { requester: requester.clone(), update }) {
             log::error!(
                 "The RX part of the {} channel is closed, but an update is received.\nError:{}\n",
                 variant,
@@ -44,28 +47,31 @@ where
 ///
 /// See the [module-level documentation](crate::dispatching) for the design
 /// overview.
-pub struct Dispatcher {
-    bot: Bot,
+pub struct Dispatcher<R> {
+    requester: R,
 
-    messages_queue: Tx<Message>,
-    edited_messages_queue: Tx<Message>,
-    channel_posts_queue: Tx<Message>,
-    edited_channel_posts_queue: Tx<Message>,
-    inline_queries_queue: Tx<InlineQuery>,
-    chosen_inline_results_queue: Tx<ChosenInlineResult>,
-    callback_queries_queue: Tx<CallbackQuery>,
-    shipping_queries_queue: Tx<ShippingQuery>,
-    pre_checkout_queries_queue: Tx<PreCheckoutQuery>,
-    polls_queue: Tx<Poll>,
-    poll_answers_queue: Tx<PollAnswer>,
+    messages_queue: Tx<R, Message>,
+    edited_messages_queue: Tx<R, Message>,
+    channel_posts_queue: Tx<R, Message>,
+    edited_channel_posts_queue: Tx<R, Message>,
+    inline_queries_queue: Tx<R, InlineQuery>,
+    chosen_inline_results_queue: Tx<R, ChosenInlineResult>,
+    callback_queries_queue: Tx<R, CallbackQuery>,
+    shipping_queries_queue: Tx<R, ShippingQuery>,
+    pre_checkout_queries_queue: Tx<R, PreCheckoutQuery>,
+    polls_queue: Tx<R, Poll>,
+    poll_answers_queue: Tx<R, PollAnswer>,
 }
 
-impl Dispatcher {
-    /// Constructs a new dispatcher with the specified `bot`.
+impl<R> Dispatcher<R>
+where
+    R: Send + 'static,
+{
+    /// Constructs a new dispatcher with the specified `requester`.
     #[must_use]
-    pub fn new(bot: Bot) -> Self {
+    pub fn new(requester: R) -> Self {
         Self {
-            bot,
+            requester,
             messages_queue: None,
             edited_messages_queue: None,
             channel_posts_queue: None,
@@ -81,10 +87,12 @@ impl Dispatcher {
     }
 
     #[must_use]
-    fn new_tx<H, Upd>(&self, h: H) -> Tx<Upd>
+    #[allow(clippy::unnecessary_wraps)]
+    fn new_tx<H, Upd>(&self, h: H) -> Tx<R, Upd>
     where
-        H: DispatcherHandler<Upd> + Send + 'static,
+        H: DispatcherHandler<R, Upd> + Send + 'static,
         Upd: Send + 'static,
+        R: Send + 'static,
     {
         let (tx, rx) = mpsc::unbounded_channel();
         tokio::spawn(async move {
@@ -97,7 +105,7 @@ impl Dispatcher {
     #[must_use]
     pub fn messages_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<Message> + 'static + Send,
+        H: DispatcherHandler<R, Message> + 'static + Send,
     {
         self.messages_queue = self.new_tx(h);
         self
@@ -106,7 +114,7 @@ impl Dispatcher {
     #[must_use]
     pub fn edited_messages_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<Message> + 'static + Send,
+        H: DispatcherHandler<R, Message> + 'static + Send,
     {
         self.edited_messages_queue = self.new_tx(h);
         self
@@ -115,7 +123,7 @@ impl Dispatcher {
     #[must_use]
     pub fn channel_posts_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<Message> + 'static + Send,
+        H: DispatcherHandler<R, Message> + 'static + Send,
     {
         self.channel_posts_queue = self.new_tx(h);
         self
@@ -124,7 +132,7 @@ impl Dispatcher {
     #[must_use]
     pub fn edited_channel_posts_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<Message> + 'static + Send,
+        H: DispatcherHandler<R, Message> + 'static + Send,
     {
         self.edited_channel_posts_queue = self.new_tx(h);
         self
@@ -133,7 +141,7 @@ impl Dispatcher {
     #[must_use]
     pub fn inline_queries_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<InlineQuery> + 'static + Send,
+        H: DispatcherHandler<R, InlineQuery> + 'static + Send,
     {
         self.inline_queries_queue = self.new_tx(h);
         self
@@ -142,7 +150,7 @@ impl Dispatcher {
     #[must_use]
     pub fn chosen_inline_results_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<ChosenInlineResult> + 'static + Send,
+        H: DispatcherHandler<R, ChosenInlineResult> + 'static + Send,
     {
         self.chosen_inline_results_queue = self.new_tx(h);
         self
@@ -151,7 +159,7 @@ impl Dispatcher {
     #[must_use]
     pub fn callback_queries_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<CallbackQuery> + 'static + Send,
+        H: DispatcherHandler<R, CallbackQuery> + 'static + Send,
     {
         self.callback_queries_queue = self.new_tx(h);
         self
@@ -160,7 +168,7 @@ impl Dispatcher {
     #[must_use]
     pub fn shipping_queries_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<ShippingQuery> + 'static + Send,
+        H: DispatcherHandler<R, ShippingQuery> + 'static + Send,
     {
         self.shipping_queries_queue = self.new_tx(h);
         self
@@ -169,7 +177,7 @@ impl Dispatcher {
     #[must_use]
     pub fn pre_checkout_queries_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<PreCheckoutQuery> + 'static + Send,
+        H: DispatcherHandler<R, PreCheckoutQuery> + 'static + Send,
     {
         self.pre_checkout_queries_queue = self.new_tx(h);
         self
@@ -178,7 +186,7 @@ impl Dispatcher {
     #[must_use]
     pub fn polls_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<Poll> + 'static + Send,
+        H: DispatcherHandler<R, Poll> + 'static + Send,
     {
         self.polls_queue = self.new_tx(h);
         self
@@ -187,7 +195,7 @@ impl Dispatcher {
     #[must_use]
     pub fn poll_answers_handler<H>(mut self, h: H) -> Self
     where
-        H: DispatcherHandler<PollAnswer> + 'static + Send,
+        H: DispatcherHandler<R, PollAnswer> + 'static + Send,
     {
         self.poll_answers_queue = self.new_tx(h);
         self
@@ -197,9 +205,13 @@ impl Dispatcher {
     ///
     /// The default parameters are a long polling update listener and log all
     /// errors produced by this listener).
-    pub async fn dispatch(&self) {
+    pub async fn dispatch(&self)
+    where
+        R: Requester + Clone,
+        <R as Requester>::GetUpdatesFaultTolerant: Send,
+    {
         self.dispatch_with_listener(
-            update_listeners::polling_default(self.bot.clone()),
+            update_listeners::polling_default(self.requester.clone()),
             LoggingErrorHandler::with_custom_text("An error from the update listener"),
         )
         .await;
@@ -215,6 +227,7 @@ impl Dispatcher {
         UListener: UpdateListener<ListenerE> + 'a,
         Eh: ErrorHandler<ListenerE> + 'a,
         ListenerE: Debug,
+        R: Requester + Clone,
     {
         let update_listener = Box::pin(update_listener);
 
@@ -235,11 +248,16 @@ impl Dispatcher {
 
                     match update.kind {
                         UpdateKind::Message(message) => {
-                            send!(&self.bot, &self.messages_queue, message, UpdateKind::Message);
+                            send!(
+                                &self.requester,
+                                &self.messages_queue,
+                                message,
+                                UpdateKind::Message
+                            );
                         }
                         UpdateKind::EditedMessage(message) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.edited_messages_queue,
                                 message,
                                 UpdateKind::EditedMessage
@@ -247,7 +265,7 @@ impl Dispatcher {
                         }
                         UpdateKind::ChannelPost(post) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.channel_posts_queue,
                                 post,
                                 UpdateKind::ChannelPost
@@ -255,7 +273,7 @@ impl Dispatcher {
                         }
                         UpdateKind::EditedChannelPost(post) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.edited_channel_posts_queue,
                                 post,
                                 UpdateKind::EditedChannelPost
@@ -263,7 +281,7 @@ impl Dispatcher {
                         }
                         UpdateKind::InlineQuery(query) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.inline_queries_queue,
                                 query,
                                 UpdateKind::InlineQuery
@@ -271,7 +289,7 @@ impl Dispatcher {
                         }
                         UpdateKind::ChosenInlineResult(result) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.chosen_inline_results_queue,
                                 result,
                                 UpdateKind::ChosenInlineResult
@@ -279,7 +297,7 @@ impl Dispatcher {
                         }
                         UpdateKind::CallbackQuery(query) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.callback_queries_queue,
                                 query,
                                 UpdateKind::CallbackQuer
@@ -287,7 +305,7 @@ impl Dispatcher {
                         }
                         UpdateKind::ShippingQuery(query) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.shipping_queries_queue,
                                 query,
                                 UpdateKind::ShippingQuery
@@ -295,18 +313,18 @@ impl Dispatcher {
                         }
                         UpdateKind::PreCheckoutQuery(query) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.pre_checkout_queries_queue,
                                 query,
                                 UpdateKind::PreCheckoutQuery
                             );
                         }
                         UpdateKind::Poll(poll) => {
-                            send!(&self.bot, &self.polls_queue, poll, UpdateKind::Poll);
+                            send!(&self.requester, &self.polls_queue, poll, UpdateKind::Poll);
                         }
                         UpdateKind::PollAnswer(answer) => {
                             send!(
-                                &self.bot,
+                                &self.requester,
                                 &self.poll_answers_queue,
                                 answer,
                                 UpdateKind::PollAnswer