+++ src/types/non_telegram_types/mod.rs
@@ -1,7 +1,0 @@
-pub use country_code::*;
-pub use currency::*;
-pub use mime_wrapper::*;
-
-mod country_code;
-mod currency;
-mod mime_wrapper;